feat: rename AssociateWithJobsModal to AssociatePayCodesWithJobsModal and implement new modal component
